[#14] _class01_show.py created, get_available_spectral_model_functions() implemented
