+++ spectrally/_class01_SpectralModel.py
@@ -7,6 +7,7 @@ import copy
 
 from ._class00_SpectralLines import SpectralLines as Previous
 from . import _class01_check_model as _check_model
+from . import _class01_show as _show
 from . import _class01_check_constraints as _check_constraints
 from . import _class01_fit_func as _fit_func
 from . import _class01_interpolate as _interpolate
@@ -101,13 +102,13 @@ class SpectralModel(Previous):
 
     def _get_show_obj(self, which=None):
         if which == self._which_model:
-            return _check_model._show
+            return _show._show
         else:
             return super()._get_show_obj(which)
 
     def _get_show_details(self, which=None):
         if which == self._which_model:
-            return _check_model._show_details
+            return _show._show_details
         else:
             super()._get_show_details(which)
